Обновление запросов локаций.

+++ src/Application/Events.Application.Services/Features/Locations/Commands/Delete/DeleteLocationHandler.cs
@@ -1,5 +1,7 @@
 ﻿using Events.Application.Services.Shared;
 using Events.Domain.Aggregates.Locations;
+using Events.Domain.Aggregates.Locations.Errors;
+using Events.Domain.Exceptions;
 using MediatR;
 
 namespace Events.Application.Services.Features.Locations.Commands.Delete;
@@ -10,6 +12,9 @@ public sealed class DeleteLocationHandler(IRepository<Location> locationReposito
     public async Task Handle(DeleteLocationCommand request, CancellationToken cancellationToken)
     {
         var location = await locationRepository.GetByIdAsync(request.LocationId, cancellationToken);
+
+        if (location == null)
+            throw new NotFoundException(LocationErrorMessages.NotFoundById(request.LocationId));
 
         await locationRepository.DeleteAsync(location, cancellationToken);
     }
